requirements.txt
torch
torchvision
torch-geometric
numpy
matplotlib
scipy
h5py
tqdm
pyyaml
